Move `createMain` to utils file
